refactor stuff to Nettoarvo component
