+++ frontend_convex_nextjs_clerk/components/dashboard/nettoarvo.tsx
@@ -1,7 +1,23 @@
 import { Card, CardHeader, CardTitle, CardContent, CardFooter } from "@/components/ui/card";
+import {
+  useQuery,
+} from "convex/react";
+import { api } from "@/convex/_generated/api";
 
-
-export default function Nettoarvo({ viewer, netValue }: { viewer: string, netValue: number }) {
+export default function Nettoarvo() {
+  const { viewer, numbers } =
+    useQuery(api.myFunctions.listNumbers, {
+      count: 5,
+    }) ?? {};
+  const netValue = useQuery(api.myFunctions.netValue);
+  
+  if (viewer === undefined || numbers === undefined) {
+    return (
+      <div className="mx-auto">
+        <p>loading... (consider a loading skeleton)</p>
+      </div>
+    );
+  }
   return (
 
     <div className="flex flex-col gap-8  w-full max-w-5xl mx-auto">
@@ -10,7 +26,7 @@ export default function Nettoarvo({ viewer, netValue }: { viewer: string, netVal
         <CardTitle>Welcome {viewer ?? "Anonymous"}!</CardTitle>
       </CardHeader>
       <CardContent>
-        <p>Nettoarvo: {netValue}</p>
+        <p>Nettoarvo: {netValue?.[0]?.value ?? 0}</p>
       </CardContent>
       <CardFooter></CardFooter>
     </Card>
